% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daedalus.compare-package.R
\docType{package}
\name{daedalus.compare-package}
\alias{daedalus.compare}
\alias{daedalus.compare-package}
\title{daedalus.compare: Run Multiple Scenarios of the Daedalus Model}
\description{
Run multiple pandemic response scenarios using the daedalus package and access epidemiological and economic outcomes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jameel-institute/daedalus.compare}
  \item Report bugs at \url{https://github.com/jameel-institute/daedalus.compare/issues}
}

}
\author{
\strong{Maintainer}: Pratik Gupte \email{pratik.gupte@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5294-7819}{ORCID})

Other contributors:
\itemize{
  \item Abdul Latif Jameel Institute for Disease and Emergency Analytics [funder]
  \item Imperial College of Science, Technology and Medicine [copyright holder, funder]
}

}
\keyword{internal}
