% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_helpers.R
\name{get_cost_data}
\alias{get_cost_data}
\title{Get cost data from DAEDALUS scenarios}
\usage{
get_cost_data(dt, disease_tags = "default")
}
\arguments{
\item{...}{}
}
\description{
Get cost data from DAEDALUS scenarios
}
