% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_helpers.R
\name{get_costs_list}
\alias{get_costs_list}
\title{Get pandemic costs from a list of model outputs}
\usage{
get_costs_list(l, names)
}
\arguments{
\item{l}{A list of \verb{<daedalus_output>} objects; each object will be passed to
\code{daedalus::get_costs()}.}

\item{names}{A character vector, intended to apply to elements of \code{l}.}
}
\description{
See \code{daedalus::get_costs()} for more information.
}
\keyword{internal}
