% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_helpers.R
\name{get_econ_cost_data}
\alias{get_econ_cost_data}
\title{Get economic cost details from an output data.table}
\usage{
get_econ_cost_data(dt)
}
\arguments{
\item{dt}{A \verb{<data.table>} resulting from \code{run_scenarios()}.}
}
\value{
A \verb{<data.frame>} summarising economic costs over the timeframe of the
modelled epidemics, broken down into costs due to restrictions and illness-
related absences.
}
\description{
Get economic cost details from an output data.table
}
