% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_helpers.R
\name{get_epicurve_data}
\alias{get_epicurve_data}
\title{Title}
\usage{
get_epicurve_data(dt, disease_tags = "default", format = c("long", "wide"))
}
\arguments{
\item{disease_tags}{}
}
\description{
Title
}
