% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_helpers.R
\name{get_epidata_list}
\alias{get_epidata_list}
\title{Get incidence from a list of model outputs}
\usage{
get_epidata_list(l, names = c("lower", "mean", "upper"), ...)
}
\arguments{
\item{...}{}
}
\description{
Get incidence from a list of model outputs
}
