% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_helpers.R
\name{get_summary_list}
\alias{get_summary_list}
\title{Get epidemic summary from a list of model outputs}
\usage{
get_summary_list(l, names = c("lower", "mean", "upper"), ...)
}
\arguments{
\item{...}{}
}
\description{
Get epidemic summary from a list of model outputs
}
