% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{NAMES_VECTOR_INF_PARAMS}
\alias{NAMES_VECTOR_INF_PARAMS}
\title{Names of age-varying infection parameters}
\format{
An object of class \code{character} of length 4.
}
\usage{
NAMES_VECTOR_INF_PARAMS
}
\description{
Names of age-varying infection parameters
}
\keyword{datasets}
