% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npi_helpers.R
\name{validate_npi_list_input}
\alias{validate_npi_list_input}
\title{Check multi-NPI inputs}
\usage{
validate_npi_list_input(x, timed_only = TRUE)
}
\arguments{
\item{x}{A list to be checked as a list of \verb{<daedalus_npi>}. Must have a
minimum length of 2 elements.}

\item{timed_only}{A boolean defaulting to \code{TRUE} for whether only
time-limited NPIs are allowed. This is typically the case for a real-time
modelling exercise.}
}
\value{
Either \code{x}, if it is a list of \verb{<daedalus_npi>}, or an error side-effect if
not.
}
\description{
A helper function to check whether a list of NPI objects has been passed as
input. This could live in \pkg{daedalus} similar to
\code{daedalus:::validate_infection_list_input()}, but lives here instead.
}
\keyword{internal}
