% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vax_helpers.R
\name{validate_vax_list_input}
\alias{validate_vax_list_input}
\title{Check multi-vaccination inputs}
\usage{
validate_vax_list_input(x)
}
\arguments{
\item{x}{A list to be checked as a list of \verb{<daedalus_vaccination>}.
Must have a minimum length of 2 elements.}
}
\value{
Either \code{x}, if it is a list of \verb{<daedalus_vaccination>}, or an error
side-effect if not.
}
\description{
A helper function to check whether a list of vaccination objects has been
passed as input.
}
\keyword{internal}
