% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_helpers.R
\name{get_costs_list}
\alias{get_costs_list}
\title{Get costs from a list of model outputs}
\usage{
get_costs_list(l, names = c("lower", "mean", "upper"), summarise_as = "domain")
}
\arguments{
\item{summarise_as}{}
}
\description{
Get costs from a list of model outputs
}
