% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_helpers.R
\name{get_costs_list}
\alias{get_costs_list}
\title{Get pandemic costs from a list of model outputs}
\usage{
get_costs_list(l, names, summarise_as = c("domain", "total"))
}
\arguments{
\item{l}{A list of \verb{<daedalus_output>} objects; each object will be passed to
\code{daedalus::get_costs()}.}

\item{names}{A character vector, intended to apply to elements of \code{l}.}

\item{summarise_as}{A string specifying whether costs are split by domain.
See \code{daedalus::get_costs()} for more information.}
}
\description{
Get pandemic costs from a list of model outputs
}
\keyword{internal}
