% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_helpers.R
\name{get_econ_cost_data}
\alias{get_econ_cost_data}
\title{Get economic cost details from an output data.table}
\usage{
get_econ_cost_data(dt)
}
\description{
Get economic cost details from an output data.table
}
