% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_helpers.R
\name{get_econ_costs_list}
\alias{get_econ_costs_list}
\title{Get details of economic costs from output with uncertainty}
\usage{
get_econ_costs_list(l)
}
\description{
Get details of economic costs from output with uncertainty
}
