% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_helpers.R
\name{get_epicurve_data}
\alias{get_epicurve_data}
\title{Get epidemiological curves from model data}
\usage{
get_epicurve_data(dt, disease_tags = "default", format = c("long", "wide"))
}
\arguments{
\item{dt}{A \verb{<data.table>} resulting from \code{run_scenarios()}.}

\item{disease_tags}{A character vector giving names for replicates within
each scenario.}

\item{format}{A string for whether the data should be returned in \code{"long"} or
\code{"wide"} format. Returning a 'wide' data.frame is only recommended when there
is a small number of \code{disease_tags}, typically a triplet of "low", "medium",
and "high" risk scenarios.}
}
\value{
A \verb{<data.frame>} of individuals in each epidemiological compartments
over the timeframe of the modelled epidemics.
}
\description{
Get epidemiological curves from model data
}
