% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_helpers.R
\name{get_epidata_list}
\alias{get_epidata_list}
\title{Get incidence data from a list of model outputs}
\usage{
get_epidata_list(l, names)
}
\arguments{
\item{l}{A list of \verb{<daedalus_output>} objects; each object will be passed to
\code{daedalus::get_costs()}.}

\item{names}{A character vector, intended to apply to elements of \code{l}.}
}
\description{
Get incidence data from a list of model outputs
}
\keyword{internal}
