% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_helpers.R
\name{get_summary_data}
\alias{get_summary_data}
\title{Get summary data from DAEDALUS scenarios}
\usage{
get_summary_data(dt, disease_tags = "default", ..., format = c("long", "wide"))
}
\arguments{
\item{...}{}
}
\description{
Get summary data from DAEDALUS scenarios
}
