% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_helpers.R
\name{get_summary_list}
\alias{get_summary_list}
\title{Get epidemic summary data from a list of model outputs}
\usage{
get_summary_list(l, names = c("lower", "mean", "upper"))
}
\arguments{
\item{l}{A list of \verb{<daedalus_output>} objects; each object will be passed to
\code{daedalus::get_costs()}.}

\item{names}{A character vector, intended to apply to elements of \code{l}.}
}
\description{
Get epidemic summary data from a list of model outputs
}
\keyword{internal}
