% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_helpers.R
\name{make_infection_samples}
\alias{make_infection_samples}
\title{Generate multiple infections with parameter distributions}
\usage{
make_infection_samples(
  name,
  samples = 100,
  param_distributions = list(r0 = distributional::dist_beta(2, 5)),
  param_ranges = list(r0 = c(1, 2))
)
}
\arguments{
\item{name}{An infection name from among \code{daedalus::epidemic_names}.}

\item{samples}{The number of samples to generate.}

\item{param_distributions}{A named list of \verb{<distribution>} class
objects provided by \pkg{distributional}, with names corresponding to
\code{daedalus::infection_parameter_names}. These are used to generate \code{samples}
draws from each distribution for the corresponding infection parameters.
Only arguments which do not vary by age are supported.}

\item{param_ranges}{An optional named list of two-element vectors, giving the
ranges to which samples drawn using \code{param_distributions} should be rescaled.
This allows easy rescaling of values drawn from bounded distributions (e.g.
the Beta distribution).
If passed, the list must have names corresponding to \code{param_distribution}
names. When empty, the generated values are used without scaling.}
}
\value{
A list of \verb{<daedalus_infection>} objects.
}
\description{
Generate multiple infections with parameter distributions
}
