% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_helpers.R
\name{run_scenarios}
\alias{run_scenarios}
\title{Run multiple DAEDALUS scenarios}
\usage{
run_scenarios(
  country,
  infection,
  response_strategy = NULL,
  time_end = 100,
  initial_state_manual = NULL
)
}
\arguments{
\item{country}{A country or territory object of class \verb{<daedalus_country>},
\strong{or} a country or territory name from those included in the package;
see \link[daedalus.data:country_names_codes]{daedalus.data::country_names}, \strong{or} a country ISO2 or ISO3 code; see
\link[daedalus.data:country_names_codes]{daedalus.data::country_codes_iso2c} and
\link[daedalus.data:country_names_codes]{daedalus.data::country_codes_iso3c}.
Country-specific data such as the community and workplace contacts, the
demography, and the distribution of the workforce into economic sectors is
automatically accessed from package data for the relevant country name if it
is passed as a string.
To override package defaults for country characteristics, pass a
\verb{<daedalus_country>} object instead. See \code{\link[daedalus:daedalus_country]{daedalus_country()}} for more.}

\item{infection}{A list of \verb{<daedalus_infection>} objects. Must have a
minimum length of 2.}

\item{response_strategy}{A time-limited response strategy specified as a
single \verb{<daedalus_npi>} created using \code{\link[daedalus:class_npi]{daedalus::daedalus_timed_npi()}},
or a list of such objects, or \code{NULL} for no response. Lists may include
\code{NULL}, which is useful when comparing response scenarios against the
no-response counterfactual.
If a list is passed, all elements must be named, or the list may have no
names, in which case synthetic names will be assigned and a message printed
to screen. Lists with some elements named and some unnamed are not accepted.}

\item{time_end}{A vector of integer-ish numbers giving the durations over
which to run scenarios. Each scenario is run for each \code{time_end}. The
intention is to be able to run response scenarios for different durations,
to be able to generate a time-series of how different costs accumulate.}

\item{initial_state_manual}{An optional \strong{named} list with the names
\code{p_infectious}, \code{p_asymptomatic}, and \code{p_immune}.
\code{p_infectious} and \code{p_asymptomatic} give the proportion of
infectious and symptomatic individuals in each age group and economic sector.
Defaults to \code{1e-6} and \code{0.0} respectively.
\code{p_immune} may be a single number in the range \verb{0.0 <= p_immune <= 1.0} or a
4-element vector in that range (the number of age groups in the model), for
the proportion of individuals in the population or in each age group that
have some pre-existing immunity to infection (reduced susceptibility). See
\strong{Details} for more.}
}
\value{
A \verb{<data.table>}, with data held in list-columns.
}
\description{
Run multiple DAEDALUS scenarios
}
