% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_helpers.R
\name{run_scenarios}
\alias{run_scenarios}
\title{Run multiple DAEDALUS scenarios}
\usage{
run_scenarios(
  country,
  disease_x,
  response = c("none", "elimination"),
  response_time_start = 0,
  response_time_end = 0,
  duration = 100,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Run multiple DAEDALUS scenarios
}
